% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiCGM.R
\name{BiCGM}
\alias{BiCGM}
\title{Biconjugate gradient method [BiCGM]
(Метод бисопряженных градиентов)}
\usage{
BiCGM(A, f, u, eps = 0.001, iterations = 10000)
}
\arguments{
\item{A}{- the original matrix of the operator equation - numeric matrix only (исходная матрица операторного уравнения - вещественная только)}

\item{f}{- bias - numeric or complex vector (вектор свободных членов вещественный или комплексный)}

\item{u}{- initial approximation of an unknown vector - numeric or complex vector (начальное приближение неизвестного вектора - вещественный или комплексный вектор)}

\item{eps}{- accuracy of calculation of the desired vector - numeric (точность вычисления искомого вектора - вещественная)}

\item{iterations}{- the upper limit on the number of iterations when the method diverges (ограничение сверху на число итераций при расхождении метода)}
}
\value{
u - unknown vector in some approximation (неизвестный вектор в некотором приближении)
}
\description{
Non-stationary iterative numerical method for solving SLAEs of the Krylov type. It is a generalization of the conjugate gradient method.
(Нестационарный итерационный численный метод решения СЛАУ крыловского типа. Является обобщением метода сопряжённых градиентов.)
}
\examples{
A <- diag(rnorm(5, 2), nrow = 5, ncol = 5)
u <- rnorm(5, 12)
f <- rnorm(5, 17)
solve(A) \%*\% f - BiCGM(A, f, u, iterations = 10000)
all.equal(solve(A) \%*\% f, BiCGM(A, f, u, iterations = 10000))
}
