% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CGM.R
\name{CGM}
\alias{CGM}
\title{Title}
\usage{
CGM(A, f, u, eps = 1e-05)
}
\arguments{
\item{A}{- the original matrix of the operator equation (исходная матрица операторного уравнения)}

\item{f}{- bias (вектор свободных членов)}

\item{u}{- initial approximation of an unknown vector (начальное приближение неизвестного вектора)}

\item{eps}{- accuracy of calculation of the desired vector (точность вычисления искомого вектора)}
}
\value{
u - unknown vector in some approximation (неизвестный вектор в некотором приближении)
}
\description{
Title
}
