% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{diag_mul_diag_torch}
\alias{diag_mul_diag_torch}
\title{Title}
\usage{
diag_mul_diag_torch(diag1_vec, diag2_vec, mode = "vec")
}
\arguments{
\item{diag1_vec}{}

\item{diag2_vec}{}

\item{mode}{}
}
\value{

}
\description{
Title
}
