% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{.circle}
\alias{.circle}
\title{Title}
\usage{
.circle(
  x,
  y,
  radius,
  nv = 100,
  border = NULL,
  col = NA,
  lty = 1,
  density = NULL,
  angle = 45,
  lwd = 0.5
)
}
\arguments{
\item{x}{}

\item{y}{}

\item{radius}{}

\item{nv}{}

\item{border}{}

\item{col}{}

\item{lty}{}

\item{density}{}

\item{angle}{}

\item{lwd}{}
}
\value{

}
\description{
Title
}
