% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{.drawComplex}
\alias{.drawComplex}
\title{Title}
\usage{
.drawComplex(lambs, mu, R, fileName = "complexPlot")
}
\arguments{
\item{lambs}{}

\item{mu}{}

\item{R}{}

\item{fileName}{}
}
\value{

}
\description{
Title
}
