% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{.mu2}
\alias{.mu2}
\title{Title}
\usage{
.mu2(z1, z2)
}
\arguments{
\item{z1}{- комплексное число, обозначающее начало отрезка спектра оператора на комплексной плоскости}

\item{z2}{- комплексное число, обозначающее конец отрезка спектра оператора на комплексной плоскости}
}
\value{
mu - центр окружности на комплексной плоскости, выпуклая оболочка спектра
}
\description{
Title
}
