% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_packages.R
\name{get_packages}
\alias{get_packages}
\title{Get Packages function}
\usage{
get_packages()
}
\value{

}
\description{
For downloading and load to projects GPU torch functions
}
\examples{
IMSSLAER::get_packages()
}
