% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMRES.R
\name{GMRES}
\alias{GMRES}
\title{Generalized minimal residual algorithm [GMRES]
(Обобщенный метод минимальных невязок)}
\usage{
GMRES(A, f, u, layers = 2, eps = 0.001, iterations = 10000)
}
\arguments{
\item{A}{- the original matrix of the operator equation - numeric 
or complex matrix (исходная матрица операторного уравнения - 
вещественная или комплексная)}

\item{f}{- bias - numeric or complex vector (вектор свободных членов 
вещественный или комплексный)}

\item{u}{- initial approximation of an unknown vector - numeric 
or complex vector (начальное приближение неизвестного вектора 
- вещественный или комплексный вектор)}

\item{layers}{- Krylov subspace order (порядок подпространства 
Крылова)}

\item{eps}{- accuracy of calculation of the desired vector - numeric 
(точность вычисления искомого вектора - вещественная)}

\item{iterations}{- the upper limit on the number of iterations when 
the method diverges (ограничение сверху на число итераций при расхождении метода)}
}
\value{
u - unknown vector in some approximation 
(неизвестный вектор в некотором приближении)
}
\description{
Non-stationary iterative numerical method for solving 
systems of linear algebraic equations. The projection of the vector 
onto the Krylov subspace of arbitrary order is used as a residual.
(Нестационарный итерационный численный метод решения систем 
линейных алгебраических уравнений. В качестве невязки используется 
проекция вектора на подпространство Крылова произвольного порядка.)
}
\examples{
A <- diag(rnorm(25, 5) + 1i * rnorm(25, 1), ncol = 25, nrow = 25)
f <- rnorm(25, 2)
u <- rnorm(25) 
print(GMRES(A, f, u, layers = 5, eps = 10e-07))
print(IMRES(A, f, u, eps = 10e-7))
}
