% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{GMSI.mu.file.history}
\alias{GMSI.mu.file.history}
\title{Title}
\usage{
GMSI.mu.file.history(
  A,
  f,
  u,
  input.file = "complexNumbeRS",
  out = F,
  output.file = "results",
  draw = F,
  plot.file = "complexPlot",
  eps = 0.001
)
}
\arguments{
\item{A}{}

\item{f}{}

\item{u}{}

\item{input.file}{}

\item{out}{}

\item{output.file}{}

\item{draw}{}

\item{plot.file}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
