% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{GMSI.mu.file}
\alias{GMSI.mu.file}
\title{Generalized method of simple iterations [GMSI]
(Обощённый метод простой итерации)}
\usage{
GMSI.mu.file(
  A,
  f,
  u,
  input.file = "complexNumbeRS",
  out = F,
  output.file = "results",
  draw = F,
  plot.file = "complexPlot",
  eps = 0.001,
  iterations = 10000
)
}
\arguments{
\item{A}{- the original matrix of the operator equation - 
numeric or complex matrix (исходная матрица операторного уравнения 
- вещественная или комплексная)}

\item{f}{- bias - numeric or complex vector (вектор 
свободных членов вещественный или комплексный)}

\item{u}{- initial approximation of an unknown vector - 
numeric or complex vector (начальное приближение неизвестного вектора 
- вещественный или комплексный вектор)}

\item{input.file}{- file with input data for spectrum points 
(файл с входными данными для точек спектра)}

\item{out}{- boolean variable, controls file output 
(переменная логического типа, управляет выводом в файл)}

\item{output.file}{- file with input data for the center of the circle 
(файл с входными данными для центра окружности)}

\item{draw}{- boolean type variable that controls the drawing of 
the complex plane of the spectrum of the operator (пременная 
логического типа, управляет выводом рисунка комплексной плоскости 
спектра оператора)}

\item{plot.file}{- operator plane image file name
(имя файла изображения комплексной плоскости оператора)}

\item{eps}{- accuracy of calculation of the desired vector - numeric 
(точность вычисления искомого вектора - вещественная)}

\item{iterations}{- the upper limit on the number of iterations when 
the method diverges (ограничение сверху на число итераций при 
расхождении метода)}
}
\value{
u - unknown vector in some approximation 
(неизвестный вектор в некотором приближении)
}
\description{
- Stationary iterative method for solving operator 
equations or systems of linear algebraic equations. The limitation 
is the absence among the values of the spectrum of the operator, 
the values of the mirror relative to the origin.
(Стационарный итерационный метод для решения операторных уравнений 
или систем линейных алгебраических уравнений. Ограничением ялвяется 
отсутствие среди значений спектра оператора, значений зеркальных 
относительно начала координат.)
}
