% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diag.R
\name{large_vec_norm_torch}
\alias{large_vec_norm_torch}
\title{Title}
\usage{
large_vec_norm_torch(vec, mode = "Gelder", p = 2)
}
\arguments{
\item{vec}{}

\item{mode}{}

\item{p}{}
}
\value{

}
\description{
Title
}
