% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaRemark.R
\name{MariaRemark.Alter_Johns}
\alias{MariaRemark.Alter_Johns}
\title{Alter-Johns function}
\usage{
MariaRemark.Alter_Johns(ts, p = 1)
}
\arguments{
\item{ts}{- time series of dynamic system}

\item{p}{- norm}
}
\value{

}
\description{
Alter-Johns function
}
