% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaRemark.R
\name{MariaRemark.out_of_trend}
\alias{MariaRemark.out_of_trend}
\title{Out of trend Dynamics decomposition}
\usage{
MariaRemark.out_of_trend(
  ts,
  delta,
  mode = c("AP", "GP_log", "GP_lin", "W", "HP")
)
}
\arguments{
\item{ts}{- vector of numeric data, charaterises the dynamic process}

\item{delta}{- time delta for window kernel}

\item{mode}{- type of kernel substitution}
}
\value{

}
\description{
Out of trend Dynamics decomposition
}
