% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MariaRemark.R
\name{MariaRemark.Shift_function}
\alias{MariaRemark.Shift_function}
\title{Alter-Johns general function (shift function)}
\usage{
MariaRemark.Shift_function(ts, delta_bounds = c(1, 100), mode = "AP", p = 1)
}
\arguments{
\item{ts}{- time series of dynamic process}

\item{delta_bounds}{- dt_min and dt_max of ts process}

\item{mode}{- type of kernel substitution}

\item{p}{- norm mode}
}
\value{

}
\description{
Alter-Johns general function (shift function)
}
