% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SIM_torch.R
\name{SIM_torch}
\alias{SIM_torch}
\title{Simple Iterations Method {torch}}
\usage{
SIM_torch(A, u, f, eps = 0.001, iterations = 1000)
}
\arguments{
\item{A}{}

\item{u}{}

\item{f}{}

\item{eps}{}

\item{iterations}{}
}
\value{

}
\description{
Simple Iterations Method {torch}
}
