% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIM.R
\name{Chebishev.history}
\alias{Chebishev.history}
\title{Title}
\usage{
Chebishev.history(A, f, u, lambs, layers, eps = 0.001)
}
\arguments{
\item{A}{}

\item{f}{}

\item{u}{}

\item{lambs}{}

\item{layers}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
