% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{GMSI.mu.file.history}
\alias{GMSI.mu.file.history}
\title{Title}
\usage{
GMSI.mu.file.history(
  A,
  f,
  u,
  input.file = "complexNumbeRS",
  out = F,
  output.file = "results",
  draw = F,
  plot.file = "complexPlot",
  eps = 0.001
)
}
\arguments{
\item{A}{}

\item{f}{}

\item{u}{}

\item{input.file}{}

\item{out}{}

\item{output.file}{}

\item{draw}{}

\item{plot.file}{}

\item{eps}{}
}
\value{
result - list: 
num.iter - number of iterations (число итераций); 
var - unknown vector result (результат вычисления неизвестного вектора); 
var.hist - history of computing an unknown vector (история вычисления неизвестного вектора); 
systime.iter - system time calculation (системное время вычисления);
}
\description{
Title
}
\details{
This method is necessary to preserve the history of sequential calculation of an unknown vector in order to visualize the convergence of the method 
(Данный метод необходим для сохранения истории последовательного вычисления неизвестного вектора с целью визуализации сходимости метода)
}
