% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{GMSI.mu.history}
\alias{GMSI.mu.history}
\title{Title}
\usage{
GMSI.mu.history(A, f, u, lambs, eps = 0.001)
}
\arguments{
\item{A}{}

\item{f}{}

\item{u}{}

\item{lambs}{}

\item{eps}{}
}
\value{

}
\description{
Title
}
