% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GMSI.R
\name{muFind.File}
\alias{muFind.File}
\title{Title}
\usage{
muFind.File(
  input.file = "complexNumbeRS",
  output.file = "results",
  out = T,
  draw = TRUE,
  plot.file = "complexPlot"
)
}
\arguments{
\item{input.file}{}

\item{output.file}{}

\item{out}{}

\item{draw}{}

\item{plot.file}{}
}
\value{

}
\description{
Title
}
